backend/requirements.txt
fastapi
uvicorn
pdfminer.six
langchain
elasticsearch
python-dotenv
